apps/ml/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
xgboost==2.1.0
scikit-learn==1.5.0
pandas==2.2.0
numpy==1.26.0
supabase==2.5.0
modal==0.64.0
python-dotenv==1.0.0
pydantic==2.7.0
